requirements.txt
django
djangorestframework
requests
markdown
django-filter
pytest
pytest-django
django-cors-headers
